requirements.txt
streamlit
pandas
numpy
umap-learn
plotly
matplotlib
seaborn
scikit-learn